move CacheDirectoryProvider to okhttp module
