+++ pluto-plugins/plugins/network/okhttp-network/lib/src/main/kotlin/com/pluto/plugins/network/internal/CacheDirectoryProvider.kt
@@ -2,7 +2,7 @@
  * Source: chucker - https://github.com/ChuckerTeam/chucker.git
  * License: https://github.com/ChuckerTeam/chucker/blob/develop/LICENSE.txt
  */
-package com.pluto.plugins.network.internal.interceptor.logic.core
+package com.pluto.plugins.network.internal
 
 import java.io.File
 
@@ -10,7 +10,6 @@ import java.io.File
  * An interface that returns a reference to a cache directory where temporary files can be
  * saved.
  */
-// todo move this to okhttp-network module
-fun interface CacheDirectoryProvider {
+internal fun interface CacheDirectoryProvider {
     fun provide(): File?
 }
